backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "RAG API using FastAPI and LlamaIndex"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12,<3.14"
fastapi = "0.115.9"
uvicorn = "^0.24.0"
python-dotenv = "^1.0.0"
llama-index = "^0.12.27"
llama-index-core = "^0.12.27"
openai = "^1.3.0"
setuptools = "^69.1.1"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
lxml = "^5.3.1"
colorama = "^0.4.6"
motor = "^3.3.2"
beanie = "^1.25.0"
llama-index-llms-anthropic = "^0.6.10"
llama-index-storage-index-store-mongodb = "^0.4.0"
llama-index-embeddings-voyageai = "^0.3.5"
llama-index-vector-stores-chroma = "^0.4.1"
tqdm = "^4.66.1"
llama-index-postprocessor-voyageai-rerank = "^0.3.2"
arize-phoenix-otel = "^0.9.2"
openinference-instrumentation-llama-index = "^4.2.1"
chromadb = "^1.0.12"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
serve = "scripts.serve:main"
parse-xml = "app.parsers.cli:run"
generate-summaries = "scripts.generate_summaries:main"
generate-parent-summaries = "scripts.generate_parent_summaries:main"
create-vector-index = "scripts.create_vector_index:main"
populate-embeddings = "scripts.populate_embeddings:main"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3